Create modinverse.txt
Given ax mod b = 1, and gcd(a,b) = 1, compute x, or a^-1 mod b. Example: given a = 4 and b = 7, return 2, since 4 * 2 mod 7 = 1. To compute the modulo inverse, implement the Extended Euclidean Algorithm.  Input a is in r0 and input b is in r1; output should be in r2. Inputs a and b are restricted to integers in range (1 ≤ x ≤ 127).